+++ packages/cli/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `options.whitelist` in PapiConfig to globally set the whitelist path.
 - `--skip-whitelist` CLI option to skip using a whitelist.
+- Export `WhitelistEntriesByChain` type from descriptors to help build a chain-based whitelist.
 
 ## 0.16.1 to 0.16.3 - 2025-11-27